# [on hold] Switch `convert_2q_block_matrix.rs` to use `matmul` from `faer` and an explicit version of `kron`