--- a/crates/accelerate/src/convert_2q_block_matrix.rs
+++ b/crates/accelerate/src/convert_2q_block_matrix.rs
@@ -15,13 +15,16 @@ use pyo3::wrap_pyfunction;
 use pyo3::Python;
 
 use num_complex::Complex64;
-use numpy::ndarray::{aview2, Array2, Array, ArrayView2};
+use numpy::ndarray::aview2;
 use numpy::{IntoPyArray, PyArray2, PyReadonlyArray2};
 use smallvec::SmallVec;
 
+use faer::modules::core::kron;
 use faer::modules::core::mul::matmul;
-use faer::modules::core::kron as kron;
-use faer::{Mat, mat, Parallelism};
+use faer::IntoFaerComplex;
+use faer::IntoNdarrayComplex;
+use faer::{Mat, Parallelism};
+use faer_core::c64;
 use faer_core::permutation::swap_rows;
 
 static ONE_QUBIT_IDENTITY: [[Complex64; 2]; 2] = [
@@ -36,42 +39,53 @@ pub fn blocks_to_matrix(
     py: Python,
     op_list: Vec<(PyReadonlyArray2<Complex64>, SmallVec<[u8; 2]>)>,
 ) -> PyResult<Py<PyArray2<Complex64>>> {
-    let identity = Mat::from_fn(2, 2, |i, j| ONE_QUBIT_IDENTITY[i][j] as Complex64);
-    let input_matrix = Mat::from_fn(2, 2, |i, j|  op_list[0].0.as_array()[(i,j)] as Complex64);
-    let mut matrix: Mat::<Complex64>;
-    let mut matrix_result: Mat::<Complex64>;
+    let identity = aview2(&ONE_QUBIT_IDENTITY).into_faer_complex();
+    let input_matrix = op_list[0].0.as_array().into_faer_complex();
+    let mut matrix: Mat<c64> = Mat::<c64>::zeros(4, 4);
+    let mut result: Mat<c64> = Mat::<c64>::zeros(4, 4);
 
     match op_list[0].1.as_slice() {
         [0] => kron(matrix.as_mut(), identity.as_ref(), input_matrix.as_ref()),
         [1] => kron(matrix.as_mut(), input_matrix.as_ref(), identity.as_ref()),
         [0, 1] => matrix = input_matrix.to_owned(),
-        [1, 0] => matrix = change_basis(input_matrix),
-        [] => matrix = Mat::<Complex64>::identity(4, 4),
+        [1, 0] => matrix = change_basis(input_matrix.to_owned()),
+        [] => matrix = Mat::<c64>::identity(4, 4),
         _ => unreachable!(),
     };
     for (op_matrix, q_list) in op_list.into_iter().skip(1) {
-        let op_matrix = Mat::from_fn(2, 2, |i, j| op_matrix.as_array()[(i,j)] as Complex64);
+        let op_matrix = op_matrix.as_array().into_faer_complex();
+        let mut op_result: Mat<c64> = Mat::<c64>::zeros(4, 4);
 
-        let mut result: Mat::<Complex64>;
         match q_list.as_slice() {
-            [0] => kron(result.as_mut(), identity.as_ref(), op_matrix.as_ref()),
-            [1] => kron(result.as_mut(), op_matrix.as_ref(), identity.as_ref()),
-            [1, 0] => result = change_basis(op_matrix),
-            [] => result = Mat::<Complex64>::identity(4, 4),
-            _ => result = op_matrix,
+            [0] => kron(op_result.as_mut(), identity.as_ref(), op_matrix.as_ref()),
+            [1] => kron(op_result.as_mut(), op_matrix.as_ref(), identity.as_ref()),
+            [1, 0] => op_result = change_basis(op_matrix.to_owned()),
+            [] => op_result = Mat::<c64>::identity(4, 4),
+            _ => op_result = op_matrix.to_owned(),
         };
 
-        matmul(matrix_result.as_mut(), result.as_ref(), matrix.as_ref(), None, Complex64::new(1.,0.), Parallelism::None);
+        matmul(
+            result.as_mut(),
+            op_result.as_ref(),
+            matrix.as_ref(),
+            None,
+            c64::new(1., 0.),
+            Parallelism::None,
+        );
     }
 
-    let mut array_result = Array2::from_shape_fn((4,4), |(i,j)| Complex64::new(*matrix_result.get(i, j).re, *matrix_result.get(i, j).im)); 
-    Ok(array_result.into_pyarray_bound(py).unbind())
+    Ok(result
+        .as_ref()
+        .into_ndarray_complex()
+        .to_owned()
+        .into_pyarray_bound(py)
+        .unbind())
 }
 
 /// Switches the order of qubits in a two qubit operation.
 #[inline]
-pub fn change_basis(matrix: Mat::<Complex64>) -> Mat::<Complex64> {
-    let mut trans_matrix: Mat::<Complex64> = matrix.transpose().to_owned();
+pub fn change_basis(matrix: Mat<c64>) -> Mat<c64> {
+    let mut trans_matrix: Mat<c64> = matrix.transpose().to_owned();
     swap_rows(trans_matrix.as_mut(), 1, 2);
     trans_matrix = trans_matrix.transpose().to_owned();
     swap_rows(trans_matrix.as_mut(), 1, 2);