--- a/crates/accelerate/src/convert_2q_block_matrix.rs
+++ b/crates/accelerate/src/convert_2q_block_matrix.rs
@@ -82,15 +82,17 @@ pub fn blocks_to_matrix(
 }
 
 /// Switches the order of qubits in a two qubit operation.
+/// This function will substitue `change_basis` once the
+/// `two_qubit_decompose.rs` uses Mat<c64> instead of ArrayView2
 #[inline]
 pub fn change_basis_faer(matrix: Mat<c64>) -> Mat<c64> {
     let mut trans_matrix: Mat<c64> = matrix.transpose().to_owned();
-    let (a, b) = trans_matrix.as_mut().two_rows_mut(1, 2);
-    swap_rows(a, b);
+    let (row1, row2) = trans_matrix.as_mut().two_rows_mut(1, 2);
+    swap_rows(row1, row2);
 
     trans_matrix = trans_matrix.transpose().to_owned();
-    let (a, b) = trans_matrix.as_mut().two_rows_mut(1, 2);
-    swap_rows(a, b);
+    let (row1, row2) = trans_matrix.as_mut().two_rows_mut(1, 2);
+    swap_rows(row1, row2);
 
     trans_matrix
 }