Part of #12120

This PR switches over the `convert_2q_block_matrix.rs` script to use `matmul` from `faer` instead of `dot` from `ndarray` to perform matrix multiplications. It turns out that using `faer` for matrices of size 4x4 or smaller gives us a better performance compared to `ndarray` as we can find in this [comment](https://github.com/Qiskit/qiskit/issues/12120#issuecomment-2050030794).

Summary of the changes in this PR:

- The use of an explicit/unrolled version of `kron` 2x2 that improves the performances of the `kron` calls
- The use of `matmul()` from `faer` improving the performance of the `dot()` calls
- Introduction of `common.rs` as a module with different common functions between different scripts

In a follow-up, the same change will be done for the `two_qubit_decompose.rs` script, which is still using `ndarray`, in the meantime, and working with the old `change_basis` function.